requirements.txt
# Core Flask dependencies
Flask==3.0.0
Werkzeug==3.0.1

# MongoDB driver
pymongo==4.6.1
dnspython==2.4.2

# Authentication and security
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.2
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Rate limiting
Flask-Limiter==3.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# File handling
python-magic-bin==0.4.14; platform_system == "Windows"

# Utilities
itsdangerous==2.1.2
